refactor: mobile size detect
