+++ src/components/dashboard/FaceVerification/screens/IntroScreen.js
@@ -20,7 +20,7 @@ import UserStorage from '../../../../lib/gundb/UserStorage'
 import GDStore from '../../../../lib/undux/GDStore'
 import logger from '../../../../lib/logger/pino-logger'
 import { getFirstWord } from '../../../../lib/utils/getFirstWord'
-import { getDesignRelativeHeight, getDesignRelativeWidth } from '../../../../lib/utils/sizes'
+import { getDesignRelativeHeight, getDesignRelativeWidth, isLargeDevice } from '../../../../lib/utils/sizes'
 import { withStyles } from '../../../../lib/styles'
 import { isBrowser, isE2ERunning, isIOSWeb, isMobileSafari } from '../../../../lib/utils/platform'
 import { openLink } from '../../../../lib/utils/linking'
@@ -28,7 +28,6 @@ import Config from '../../../../config/config'
 import { Permissions } from '../../../permissions/types'
 import { showQueueDialog } from '../../../common/dialogs/showQueueDialog'
 import { fireEvent, FV_CAMERAPERMISSION, FV_CANTACCESSCAMERA, FV_INTRO } from '../../../../lib/analytics/analytics'
-import { isLargeDevice } from '../../../../lib/utils/mobileSizeDetect'
 
 // assets
 import wait24hourIllustration from '../../../../assets/Claim/wait24Hour.svg'
